device_call: address udp transceiver review findings
Factor the transport-independent v1 ring-slot RPC protocol (frame lifecycle,
round-robin slot correlation, fire-and-forget response draining, TX
back-pressure, flag handshakes, key=value argument parsing) out of
CpuRoceChannel and UdpChannel into a shared RingSlotChannel base; the
channels keep only transport setup/teardown and their wait posture
(busy-spin for RDMA, short sleep for the loopback stand-in). This removes
~200 duplicated lines and the divergences the copies had already grown.

Behavioral fixes that ride along for both channels:
- A TX slot still published after the back-pressure timeout now fails the
  dispatch with DeviceCallStatus::Timeout instead of silently overwriting a
  buffer the transport may still be transmitting (previous UDP behavior
  could put a torn datagram on the wire and drop the in-flight request
  without any error).
- Port arguments (udp-port, rendezvous-port) are validated: non-numeric or
  out-of-range values raise DeviceCallError(InvalidArgument) instead of
  escaping as raw std::stoul exceptions or silently truncating through the
  uint16_t cast.

Transport fixes:
- udp_wrapper.h no longer claims the full-stride datagram mirrors the RoCE
  TX SGE (RoCE sends cu_frame_size, not the slot stride); the same-page_size
  requirement is now stated on its own terms.
- A stride mismatch between endpoints previously dropped every request with
  no diagnostic anywhere; the RX loop now warns once on stderr with both
  sizes.
- The UdpChannel header documents the v1 positional-correlation limitation
  (a lost or >timeout-late datagram permanently desyncs the channel).

Build fixes:
- install(EXPORT cudaq-realtime-targets) and the package config are no
  longer gated on CUDA_FOUND: the always-built UDP transport joins that
  export set, and CMake silently installs an orphaned archive when a target
  joins an export set that is never installed. The config file's CUDAToolkit
  dependency is now conditional on the installed build having had CUDA.
- realtime/unittests/cpu_transport comment updated: the UDP tests run before
  the libibverbs gate.

Verified: main build links both channels; standalone realtime build (CUDA
and CUDA-less) passes all 7 udp transceiver tests; a loopback harness runs
registry-created UdpChannel round trips including fire-and-forget drain and
the new port validation; the host-dispatch no-GPU tests still pass.
